LeetCode May 29

+++ src/main/java/com/ss/leetcode/LC2022/may/MaximumTotalImportanceOfRoads.java
@@ -6,7 +6,7 @@ import java.util.Comparator;
 public class MaximumTotalImportanceOfRoads {
     // https://leetcode.com/problems/maximum-total-importance-of-roads/
     public long maximumImportance(int n, int[][] roads) {
-        long[][] cities = new long[n][3];
+        int[][] cities = new int[n][3];
         for(int i = 0; i < cities.length; i++) {
             cities[i][0] = i;
         }
@@ -14,11 +14,11 @@ public class MaximumTotalImportanceOfRoads {
             cities[road[0]][1]++;
             cities[road[1]][1]++;
         }
-        Arrays.sort(cities, Comparator.comparingLong(a -> a[1]));
+        Arrays.sort(cities, Comparator.comparingInt(a -> a[1]));
         for (int i = cities.length -1; i >= 0; i--) {
             cities[i][2] = i + 1;
         }
-        Arrays.sort(cities, Comparator.comparingLong(a -> a[0]));
+        Arrays.sort(cities, Comparator.comparingInt(a -> a[0]));
         long totalImportance = 0;
         for(int[] road : roads) {
             totalImportance += cities[road[0]][2] + cities[road[1]][2];
